backend/vital_voices_project/requirements.txt
Django==4.2
django-cors-headers==4.4.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.15.2
pymssql==2.3.0
pyodbc==5.1.0
gunicorn>=20.1.0